requirements.txt
dnspython
python-whois
python-nmap
python-dotenv
